refactor: simplify ReadingCard hover state with minimalist design
- Replace complex 993-line ribbon animation with clean 221-line overlay
- Remove 10+ animated layers, glass morphism, and gradient effects
- Implement simple dark overlay that fades in on hover
- Reduce status indicators to basic colored dots
- Improve performance by eliminating complex transforms and willChange
- Create backup of original implementation (ReadingCard.v1.tsx)
- Update tests to match new simplified structure

Benefits:
- 77% reduction in code size
- Better performance with only 2 animating properties
- Cleaner UX with non-intrusive hover states
- Improved mobile experience
